requirements.txt
async-timeout==4.0.2
black==23.3.0
build==0.10.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dj-database-url==2.0.0
Django==4.2.1
django-celery-beat==2.5.0
django-crispy-forms==2.0
djangorestframework-simplejwt==5.2.2
django-debug-toolbar==4.1.0
drf-spectacular==0.26.3
filelock==3.12.0
flower==2.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pygame==2.3.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
six==1.16.0
virtualenv==20.23.0
virtualenv-clone==0.5.7


djangorestframework~=3.14.0
telegram~=0.0.1